fix: allow multiple runs of the command in the same job (#234)
* fix: allow multiple runs of the command in the same job

* test: debugging

* test: debugging

* test: adding setup-remote-docker

* test: removed setup-remote-docker

* fix: updated executors

* test: removed dlc

* fix: edited parameter

* test: removing tags for public registry

* test: removing tags parameter

* test: removing tags parameter

* test: debugging

* test: debugging

* fix: updated cimg/node version

* fix: made tag all lower case

* fix: made tag all lower case

* tesT: debugging

* test: debugging

* fix: removed tag from image

* fix: added current alias

* test: debugging

* fix: added current alias

* test: debugging

+++ src/scripts/docker-buildx.sh
@@ -53,6 +53,7 @@ if [ "${PARAM_SKIP_WHEN_TAGS_EXIST}" = "0" ] || [[ "${PARAM_SKIP_WHEN_TAGS_EXIST
   fi
 
   if [ "${PARAM_PUBLIC_REGISTRY}" == "1" ]; then
+    echo "docker tag args: ${docker_tag_args}" >> test.txt
     docker buildx build \
       -f "${PARAM_PATH}"/"${PARAM_DOCKERFILE}" \
       ${docker_tag_args} \
@@ -61,9 +62,16 @@ if [ "${PARAM_SKIP_WHEN_TAGS_EXIST}" = "0" ] || [[ "${PARAM_SKIP_WHEN_TAGS_EXIST
       "$@" \
       "${PARAM_PATH}"
   else
-    docker context create builder
-    docker run --privileged --rm tonistiigi/binfmt --install all
-    docker --context builder buildx create --use
+
+    if ! docker context ls | grep builder; then
+      # We need to skip the creation of the builder context if it's already present
+      # otherwise the command will fail when called more than once in the same job.
+
+      docker context create builder
+      docker run --privileged --rm tonistiigi/binfmt --install all
+      docker --context builder buildx create --use
+    fi
+
     docker --context builder buildx build \
       -f "${PARAM_PATH}"/"${PARAM_DOCKERFILE}" \
       ${docker_tag_args} \
